Simplify comparison of schema
