+++ packages/dds/tree/src/core/schema-stored/schema.ts
@@ -158,7 +158,7 @@ export interface TreeFieldStoredSchema {
  *
  * 2. The schema used for out of schema fields (which thus must be empty/not exist) on object and leaf nodes.
  */
-export const forbiddenFieldKindIdentifier = "Forbidden";
+export const forbiddenFieldKindIdentifier: FieldKindIdentifier = brand("Forbidden");
 
 /**
  * A schema for empty fields (fields which must always be empty).
@@ -166,7 +166,7 @@ export const forbiddenFieldKindIdentifier = "Forbidden";
  */
 export const storedEmptyFieldSchema: TreeFieldStoredSchema = {
 	// This kind requires the field to be empty.
-	kind: brand(forbiddenFieldKindIdentifier),
+	kind: forbiddenFieldKindIdentifier,
 	// This type set also forces the field to be empty not not allowing any types as all.
 	types: new Set(),
 	persistedMetadata: undefined,
