+++ packages/dds/tree/src/simple-tree/api/schemaCompatibilityTester.ts
@@ -3,27 +3,13 @@
  * Licensed under the MIT License.
  */
 
-import { assert, unreachableCase } from "@fluidframework/core-utils/internal";
-
 import type { TreeStoredSchema } from "../../core/index.js";
-import {
-	allowsRepoSuperset,
-	defaultSchemaPolicy,
-	FieldKinds,
-	isNeverTree,
-} from "../../feature-libraries/index.js";
+import { allowsRepoSuperset, defaultSchemaPolicy } from "../../feature-libraries/index.js";
 
 import type { SchemaCompatibilityStatus } from "./tree.js";
-import {
-	comparePosetElements,
-	fieldRealizer,
-	getAllowedContentDiscrepancies,
-	PosetComparisonResult,
-	type FieldDiscrepancy,
-} from "../discrepancies.js";
+import { getAllowedContentDiscrepancies } from "./discrepancies.js";
 import { toStoredSchema } from "../toStoredSchema.js";
 import type { TreeSchema } from "./configuration.js";
-import { isObjectNodeSchema } from "../node-kinds/index.js";
 
 /**
  * A collection of View information for schema
@@ -59,147 +45,29 @@ export class SchemaCompatibilityTester {
 		// In the future, this would also include things like:
 		// - fields with more allowed types in the stored schema than in the view schema have out-of-schema "unknown content" adapters
 		let canView = true;
-		// View schema allows a superset of documents that stored schema does, hence the document could be upgraded to use a persisted version
-		// of this view schema as its stored schema.
-		let canUpgrade = true;
 
-		const updateCompatibilityFromFieldDiscrepancy = (discrepancy: FieldDiscrepancy): void => {
-			switch (discrepancy.mismatch) {
-				case "allowedTypes": {
-					// Since we only track the symmetric difference between the allowed types in the view and
-					// stored schemas, it's sufficient to check if any extra allowed types still exist in the
-					// stored schema.
-					if (
-						discrepancy.stored.some(
-							(identifier) => !isNeverTree(policy, stored, stored.nodeSchema.get(identifier)),
-						)
-					) {
-						// Stored schema has extra allowed types that the view schema does not.
-						canUpgrade = false;
-						canView = false;
-					}
-
-					if (discrepancy.view.length > 0) {
-						// View schema has extra allowed types that the stored schema does not.
-						canView = false;
-					}
-					break;
-				}
-				case "fieldKind": {
-					const result = comparePosetElements(
-						discrepancy.stored,
-						discrepancy.view,
-						fieldRealizer,
-					);
-
-					if (result === PosetComparisonResult.Greater) {
-						// Stored schema is more relaxed than view schema.
-						canUpgrade = false;
-						// Get the node this field is under, if any:
-						const treeNodeSchema =
-							discrepancy.identifier === undefined
-								? undefined
-								: this.viewSchema.definitions.get(discrepancy.identifier);
-						if (
-							discrepancy.view === FieldKinds.forbidden.identifier &&
-							treeNodeSchema !== undefined &&
-							isObjectNodeSchema(treeNodeSchema) &&
-							treeNodeSchema.allowUnknownOptionalFields === true
-						) {
-							// When the application has opted into it, we allow viewing documents which have additional
-							// optional fields in the stored schema that are not present in the view schema.
-						} else {
-							canView = false;
-						}
-					}
-
-					if (result === PosetComparisonResult.Less) {
-						// View schema is more relaxed than stored schema.
-						canView = false;
-					}
-
-					if (result === PosetComparisonResult.Incomparable) {
-						canUpgrade = false;
-						canView = false;
-					}
-
-					break;
-				}
-				case "valueSchema": {
-					canView = false;
-					canUpgrade = false;
-					break;
-				}
-				default:
-					unreachableCase(discrepancy);
-			}
-		};
-
-		for (const discrepancy of getAllowedContentDiscrepancies(this.viewSchema.root, stored)) {
-			if (!canView && !canUpgrade) {
-				break;
-			}
-
-			switch (discrepancy.mismatch) {
-				case "nodeKind": {
-					const storedNodeSchema = stored.nodeSchema.get(discrepancy.identifier);
-					// We conservatively do not allow node types to change.
-					// The only time this might be valid in the sense that the data canonically converts is converting an object node
-					// to a map node over the union of all the object fields' types.
-					if (discrepancy.stored === undefined) {
-						// View schema has added a node type that the stored schema doesn't know about.
-						// Note that all cases which trigger this should also trigger an AllowedTypeDiscrepancy (where the type is used).
-						// This means this case should be redundant and could be removed in the future if there is a reason to do so
-						// (like simplifying enablable type support).
-						// See the TODO in getAllowedContentDiscrepancies.
-						canView = false;
-					} else if (discrepancy.view === undefined) {
-						const storedIsNever =
-							storedNodeSchema !== undefined
-								? isNeverTree(policy, stored, storedNodeSchema)
-								: true;
-						if (!storedIsNever) {
-							// Stored schema has a node type that the view schema doesn't know about.
-							// The design of allowUnknownOptionalFields allows adding new optional content to types,
-							// and the new optional content can use new types.
-							// Therefore this case needs to be supported for viewing.
-							// However, it is not supported for upgrade as the stored document must be newer (or at least more general) than the view in this case.
-							// The fact that there might be removed trees with a root of this type which only appears in the stored schema is why this must set canUpgrade to false.
-							canUpgrade = false;
-						}
-					} else {
-						// Node type exists in both schemas but kind has changed. We conservatively never allow this.
-						// See note above about cases where this could be allowed if needed.
-						canView = false;
-						canUpgrade = false;
-					}
-					break;
-				}
-				case "valueSchema":
-				case "allowedTypes":
-				case "fieldKind": {
-					updateCompatibilityFromFieldDiscrepancy(discrepancy);
-					break;
-				}
-				case "fields": {
-					discrepancy.differences.forEach(updateCompatibilityFromFieldDiscrepancy);
-					break;
-				}
-				// No default
-			}
+		for (const _discrepancy of getAllowedContentDiscrepancies(this.viewSchema, stored)) {
+			canView = false;
+			break;
 		}
 
-		if (canUpgrade) {
-			assert(
-				allowsRepoSuperset(policy, stored, toStoredSchema(this.viewSchema.root)),
-				0xbf2 /* View schema must be a superset of the stored schema to allow upgrade */,
-			);
-		}
+		const canUpgrade = allowsRepoSuperset(
+			policy,
+			stored,
+			toStoredSchema(this.viewSchema.root),
+		);
+
+		// If true, then upgrading has no effect on what can be stored in the document.
+		// TODO: This should likely be changed to indicate up a schema upgrade would be a no-op, including stored schema metadata.
+		const isEquivalent =
+			canView &&
+			canUpgrade &&
+			allowsRepoSuperset(policy, toStoredSchema(this.viewSchema.root), stored);
 
 		return {
 			canView,
 			canUpgrade,
-			isEquivalent: canView && canUpgrade,
+			isEquivalent,
 		};
 	}
 }
