+++ packages/dds/tree/src/simple-tree/index.ts
@@ -249,11 +249,3 @@ export {
 	nullSchema,
 } from "./leafNodeSchema.js";
 export type { LeafSchema } from "./leafNodeSchema.js";
-export {
-	getAllowedContentDiscrepancies,
-	type FieldDiscrepancy,
-	comparePosetElements,
-	fieldRealizer,
-	type Realizer,
-	PosetComparisonResult,
-} from "./discrepancies.js";
